+++ src/components/settings/__tests__/coach-section.test.tsx
@@ -31,7 +31,7 @@ vi.mock("@/hooks/use-auth", () => ({
 }));
 
 import { I18nProvider } from "@/lib/i18n/context";
-import { SettingsSectionFrame } from "../settings-section-frame";
+import { SettingsSectionFrame } from "./section-frame-harness";
 import { CoachSection } from "../coach-section";
 
 function render(locale: "en" | "de" = "en") {
