+++ src/components/settings/__tests__/export-section.test.tsx
@@ -44,7 +44,7 @@ vi.mock("@/hooks/use-auth", () => ({
 }));
 
 import { I18nProvider } from "@/lib/i18n/context";
-import { SettingsSectionFrame } from "../settings-section-frame";
+import { SettingsSectionFrame } from "./section-frame-harness";
 import { ExportSection } from "../export-section";
 
 function render(node: React.ReactElement, locale: "en" | "de" = "en") {
