+++ src/components/settings/__tests__/medications-section.test.tsx
@@ -65,7 +65,7 @@ vi.mock("@/hooks/use-auth", () => ({
 }));
 
 import { I18nProvider } from "@/lib/i18n/context";
-import { SettingsSectionFrame } from "../settings-section-frame";
+import { SettingsSectionFrame } from "./section-frame-harness";
 import { MedicationsSection } from "../medications-section";
 
 function render(locale: "en" | "de" = "en"): string {
